% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flu289s.R
\docType{data}
\name{flu289s}
\alias{flu289s}
\title{A timetree of 289 influenza H1N1 viruses}
\format{An object of class \code{phylo} of length 4.}
\usage{
flu289s
}
\description{
An object of class phylo. The phylogeny contains H1N1 influenza viruses
isolated from birds, swine and humans. The viruses date between 1918 to
2009. Branch lengths are in units of 1 year.
}
\examples{
data(flu289s)
plot(flu289s, show.tip.label=FALSE)

}
\keyword{datasets}
