% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pri10s.R
\docType{data}
\name{pri10s}
\alias{pri10s}
\title{A timetree of 10 primate species}
\format{An object of class \code{phylo} of length 4.}
\usage{
pri10s
}
\description{
An object of class phylo. The phylogeny contains 9 primates and 1
outgroup (the tree shrew). The branch lengths are in units of 1 million
years ago.
}
\examples{
data(pri10s)
plot(pri10s)

}
\keyword{datasets}
