% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relaxed.tree.R
\name{relaxed.tree}
\alias{relaxed.tree}
\title{Simulate branch lengths on a phylogeny under a relaxed clock}
\usage{
relaxed.tree(tree, model, r, s2)
}
\arguments{
\item{tree}{an object of class phylo representing a bifurcating phylogeny}

\item{model}{character, the relaxed clock model}

\item{r}{numeric, the mean rate in substitutions per site}

\item{s2}{numeric, the rate "diffusion" parameter for the relaxed clocks}
}
\value{
An object of class phylo with branch lengths in substitutions per
  site.
}
\description{
Simulate branch lengths on a phylogeny under a relaxed clock
}
\details{
The \code{tree} is assummed to be a timetree. Thus, if all the tip
  species are extant, then \code{tree} must be ultrametric. If \code{tree} is
  not ultrametric then it is assummed you have extinct tips. The \code{tree}
  must be rooted.

  The options for \code{model} are "clk", "iln", "gbm_RY07", and "gbm0", for
  the strict clock, the independent log-normal rates, and the
  geometric-Brownian motion rates, respectively.

  If \code{model == "clk"} the branch lengths of \code{tree} are multiplied
  by \code{r}. For the other models, \eqn{n} rates (one for each branch in
  the \code{s} species phylogeny) are sampled from the appropriate
  distribution. The branch lengths in \code{tree} are then multiplied by the
  corresponding rates.

  Model "gbm_RY07" is the the one implemented by Rannala and Yang (2007) with
  the mean of the rate stabilized across the branches of the tree, i.e., this
  version of the GBM process is a martingale. Mean stabilization is achieved
  by adding a drift factor to the Brownian process on the log-rate, and thus
  the process on the log-rate is no longer a martingale. An alternative
  version "gbm0" is provided in which the drift factor is removed. In this
  version the process on the log-rate is now a martingale while the process
  on the rate is not. The drift factor causes collapsation of the rates to
  zero as time goes to infinity.
}
\examples{
require(ape)
par(mfrow=c(2,3))

data(pri10s)
# Simulate using autocorrelated log-normal rates on a primate phylogeny:
tt <- relaxed.tree(pri10s, model="gbm_RY07", r=.04e-2, s2=.26e-2)

# The relaxed tree (branch lengths are in substitutions per site):
plot(tt, main="Relaxed primate tree (subs per site)")

# The ultrametric timetree of Primates used in the simulation:
plot(pri10s, main="Primates timetree (Ma)")
axisPhylo() # Time unit in 1 Ma.

# Plot the branch lengths for both trees against each other:
plot(pri10s$edge.length, tt$edge.length,
xlab="Branch lengths (Million years)", ylab="Branch lengths (subs per site)")
abline(0, .04e-2) # the slope is the substitution rate, r

data(flu289s)
# Simulate using independent log-normal rates on an influenza H1N1 phylogeny:
tt2 <- relaxed.tree(flu289s, model="iln", r=.15e-2, s2=.45)

# The relaxed tree of influenza:
plot(tt2, show.tip.label=FALSE, main="Relaxed influenza tree (subs per site)")

# The timetree of Influenza (not ultrametric):
plot(flu289s, show.tip.label=FALSE, main="Influenza H1N1 timetree (y)")
axisPhylo(root.time=1907.35, backward=FALSE) # Time unit in 1 y.

# Plot the branch lengths:
plot(flu289s$edge.length, tt2$edge.length,
xlab="Branch lengths (years)", ylab="Branch lengths (subs per site)")
abline(0, .15e-2) # the slope is the substitution rate, r

}
\references{
Drummond et al. (2006) \emph{Relaxed phylogenetics and dating with
  confidence.} PLoS Biology, 4(5): e88.

  Panchaksaram et al. (2024) \emph{Bayesian selection of relaxed-clock
  models: Distinguishing between independent and autocorrelated rates.}
  Systematic Biology, 74: 323--334.

  Rannala and Yang (2007) \emph{Inferring speciation times under an
  episodic molecular clock.} Systematic Biology, 56: 453--466.
}
\author{
Mario dos Reis
}
