% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relaxed.tree.R
\name{correlated.trees}
\alias{correlated.trees}
\title{Simulate correlated branch lengths among loci on a phylogeny under a relaxed clock}
\usage{
correlated.trees(tree, model, r, s2, nloci, corr)
}
\arguments{
\item{tree}{an object of class phylo representing a bifurcating phylogeny}

\item{model}{character, the relaxed clock model}

\item{r}{numeric, the mean rate in substitutions per site}

\item{s2}{numeric, the rate "diffusion" parameter for the relaxed clocks}

\item{nloci}{numeric, the number of trees to simulate (one per locus)}

\item{corr, }{numeric, the correlation of log rates among loci}
}
\value{
A list with two elements: A list of length \code{nloci} of trees of
  class phylo with branch lengths in substitutions per site, and a matrix of
  branch rates for each locus.
}
\description{
Simulate correlated branch lengths among loci on a phylogeny under a relaxed clock
}
\details{
A total of \code{nloci} trees are simulated, with the log-rates
for branches across trees having correlation \code{corr}.
}
\examples{
require(ape)
par(mfrow=c(2,3))

data(pri10s)
# ILN model:
# Simulate using autocorrelated log-normal rates on a primate phylogeny,
# with no correlation among three loci:
iln0 <- correlated.trees(pri10s, model="iln", r=.04e-2, s2=.1, 3, 0)
lapply(iln0$trees, plot)
# Repeat with strong correlation among loci:
ilnc <- correlated.trees(pri10s, model="iln", r=.04e-2, s2=.1, 3, 0.9)
lapply(ilnc$trees, plot)

# GBM model:
# Simulate using autocorrelated log-normal rates on a primate phylogeny,
# with no correlation among three loci:
gbm0 <- correlated.trees(pri10s, model="gbm_RY07", r=.04e-2, s2=.26e-2, 3, 0)
lapply(gbm0$trees, plot)
# Repeat with strong correlation among loci:
gbmc <- correlated.trees(pri10s, model="gbm_RY07", r=.04e-2, s2=.26e-2, 3, 0.9)
lapply(gbmc$trees, plot)



}
\seealso{
\link{relaxed.tree} to simulate a single tree.
}
\author{
Mario dos Reis
}
