% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simclock.R
\name{relaxed.tree}
\alias{relaxed.tree}
\title{Simulate branch lengths on a phylogeny under a relaxed clock}
\usage{
relaxed.tree(tree, model, r, s2)
}
\arguments{
\item{tree}{an object of class phylo representing a bifurcating phylogeny}

\item{model}{character, the relaxed clock model}

\item{r}{numeric, the mean rate in substitutions per site}

\item{s2}{numeric, the rate "diffusion" parameter for the relaxed clocks}
}
\value{
An object of class phylo with branch lengths in substitutions per
  site.
}
\description{
Simulate branch lengths on a phylogeny under a relaxed clock
}
\details{
The \code{tree} is assummed to be a timetree. Thus, if all your tip
  species are extant, then \code{tree} must be ultrametric. If \code{tree} is
  not ultrametric then it is assummed you have extinct tips. The \code{tree}
  must be rooted and strictly bifurcating.

  The options for \code{model} are "clk", "iln", and "gbm_RY07", for the
  strict clock, the independent log-normal rates, and the geometric-Brownian
  motion rates, respectively (see Rannala and Yang, 2007). If \code{model ==
  "clk"} the branch lengths of \code{tree} are multiplied by \code{r}. If
  \code{model == "iln"} or \code{model == "gbm_RN07"}, \eqn{n = 2*s - 2}
  rates (one for each branch in the \code{s} species phylogeny) are sampled
  from the appropriate distribution. The branch lengths in \code{tree} are
  then multiplied by the corresponding rates.
}
\references{
Yang and Rannala (2007) \emph{Inferring speciation times under an
  episodic molecular clock.} Systematic Biology, 56:453-466.
}
\author{
Mario dos Reis
}
